Final Project Continued
Added to the about section- having issues with the picture. Put it in a
div and floated it, also floated the paragraph but can’t seem to move
the picture to the correct space- it is breaking the page. Put random
copy in for now, just wanted to get the layout correct. Also added the
“Work” section- I think I would like to have that be a picture so it’s
not sections of colors, i think it would look better as your scroll if
it were a picture.